fix checking

+++ src/main/java/dev/isnow/allahfinder/checker/protocol/packet/impl/MotdHandshakePacket.java
@@ -1,5 +1,6 @@
 package dev.isnow.allahfinder.checker.protocol.packet.impl;
 
+
 import dev.isnow.allahfinder.checker.connection.ConnectAtributes;
 import dev.isnow.allahfinder.checker.protocol.packet.Packet;
 import dev.isnow.allahfinder.util.PacketUtil;
@@ -10,8 +11,8 @@ import java.util.ArrayList;
 
 public class MotdHandshakePacket extends Packet {
     private final String ip;
-    private final short port;
-    public MotdHandshakePacket(DataOutputStream outputStream, ArrayList<ConnectAtributes> connectAtributes, String ip, short port) {
+    private final int port;
+    public MotdHandshakePacket(DataOutputStream outputStream, ArrayList<ConnectAtributes> connectAtributes, String ip, int port) {
         super("MotdHandshake", outputStream, connectAtributes);
         this.ip = ip;
         this.port = port;
@@ -22,7 +23,7 @@ public class MotdHandshakePacket extends Packet {
         PacketUtil.writeByte(outputStream, 0x00);
         PacketUtil.writeVarInt(outputStream, 4);
         PacketUtil.writeString(outputStream, ip);
-        PacketUtil.writeShort(outputStream, port);
+        PacketUtil.writeShort(outputStream, (short) port);
         PacketUtil.writeVarInt(outputStream, 1);
     }
 }
